update version 1.3.3

+++ database/migrations/2024_06_22_133203_add_notification_preferences_to_users_table.php
@@ -8,25 +8,30 @@ return new class extends Migration {
 
     public function up()
     {
-        Schema::table( 'users', function ( Blueprint $table ) {
+        Schema::table( config( 'iracode-filament-notification.table', 'users' ), function ( Blueprint $table ) {
             $table->boolean( 'prefers_telegram' )->default( false );
             $table->boolean( 'prefers_sms' )->default( false );
             $table->boolean( 'prefers_bale' )->default( false );
             $table->string( 'telegram_chat_id', 30 )->nullable();
-            $table->string( 'phone', 20 )->nullable();
+            if ( config( 'iracode-filament-notification.has_column_phone', false ) )
+                $table->string( 'phone', 20 )->nullable();
             $table->string( 'bale_chat_id', 30 )->nullable();
         } );
     }
 
     public function down()
     {
-        Schema::table( 'users', function ( Blueprint $table ) {
+        Schema::table( config( 'iracode-filament-notification.table', 'users' ), function ( Blueprint $table ) {
+
             $table->dropColumn( 'prefers_telegram' );
             $table->dropColumn( 'prefers_sms' );
             $table->dropColumn( 'prefers_bale' );
+
             $table->dropColumn( 'telegram_chat_id' );
-            $table->dropColumn( 'phone' );
             $table->dropColumn( 'bale_chat_id' );
+            if ( config( 'iracode-filament-notification.has_column_phone', false ) )
+                $table->dropColumn( 'phone' );
+
         } );
     }
